Fixed user in memcached_debian.j2 template.
templates/memcached_debian.j2:
- Debian package uses the user "nobody", while Ubuntu uses the user
"memcache". The issue here was that the user was hardcoded to
"memcache" in our template and led me to think that the package was
not correctly setting up the default user (which would be *very*
weird in a Debian package). As such, the user is now set to
"nobody" or "memcache", according to "ansible_distribution".

tasks/main.yml:
- Removed tasks for user and group creation, since those are dealt
by the distibution packages.